+++ src/store/assets/index.js
@@ -6,8 +6,10 @@ export const initState = () => ({
   error: null,
   isLoading: false,
   isHandling: false,
+  isInvalid: true,
   list: [],
   detail: null,
+  newCW20: null,
 });
 
 export default {