+++ src/store/assets/mutations.js
@@ -19,4 +19,11 @@ export default {
   setDetail(state, payload) {
     state.detail = payload;
   },
+  addPropToNewCW20(state, payload) {
+    const cw20 = Object.assign({}, state.newCW20, payload);
+    state.newCW20 = cw20;
+  },
+  setIsInvalid(state, payload) {
+    state.isInvalid = payload;
+  },
 };